fix type

+++ utils/evaluate.py
@@ -48,7 +48,7 @@ def eval(valid_dataloader, model, norm_flag):
         prob_list = np.append(prob_list, avg_single_video_prob)
         label_list = np.append(label_list, avg_single_video_label)
         # compute loss and acc for every video
-        avg_single_video_output = (sum(output_dict_tmp[key]) / len(output_dict_tmp[key])).long()
+        avg_single_video_output = (sum(output_dict_tmp[key]) / len(output_dict_tmp[key]))
         avg_single_video_target = (sum(target_dict_tmp[key]) / len(target_dict_tmp[key])).long()
         print('output_dict_tmp key value: ', output_dict_tmp[key])
         print('avg_single_video_output: ', avg_single_video_output)
